chore(wmg): move Chart
