+++ frontend/src/views/WheresMyGeneV2/components/GeneSearchBar/components/SaveExport/csvUtils.ts
@@ -13,7 +13,7 @@ import { generateAndCopyShareUrl } from "../ShareButton/utils";
 
 import { deserializeCellTypeMetadata } from "src/views/WheresMyGene/components/HeatMap/utils";
 import { ChartProps } from "src/views/WheresMyGene/common/types";
-import { ChartFormat } from "src/views/WheresMyGene/components/HeatMap/components/Chart/components/Chart/hooks/utils";
+import { ChartFormat } from "src/views/WheresMyGeneV2/components/HeatMap/components/Chart/types";
 
 const NO_SELECTION_STRING = "No selection";
