+++ frontend/src/views/WheresMyGeneV2/components/HeatMap/components/Chart/types.ts
